requirements.txt
# Requires Python 3.12.

# -------------------------------------------------------
# Core data
# -------------------------------------------------------
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.8.0

# -------------------------------------------------------
# Deep learning
# -------------------------------------------------------
torch==2.6.0

# -------------------------------------------------------
# Quantum computing
# -------------------------------------------------------
pennylane==0.44.0
pennylane-qiskit==0.44.0   # PennyLane plugin for Qiskit/IBM devices

# IBM Quantum runtime (replaces deprecated qiskit-ibmq-provider)
qiskit==1.1.0
qiskit-ibm-runtime==0.23.0

# -------------------------------------------------------
# Visualisation
# -------------------------------------------------------
matplotlib==3.9.0
seaborn==0.13.2

# -------------------------------------------------------
# Utilities
# -------------------------------------------------------
python-dotenv==1.0.1        # load IBM_QUANTUM_TOKEN from .env
tqdm==4.66.4                # progress bars in training loop

# -------------------------------------------------------
# Jupyter
# -------------------------------------------------------
jupyter==1.0.0
ipykernel==6.29.4
